requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
frozenlist==1.3.3
geographiclib==1.52
geopy==2.3.0
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
magic-filter==1.0.9
MarkupSafe==2.1.2
matplotlib==3.6.3
multidict==6.0.4
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
phonenumbers==8.13.8
Pillow==9.4.0
protobuf==3.20.1
pydantic==1.10.7
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.3
telegram==0.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2

Pyrogram~=2.0.106
asyncmy
tgcrypto